fix(doc):  使用 setRenderKey 刷新 iframe
